apps/api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
python-dotenv==1.0.1
httpx==0.27.0
openai==1.21.0
python-multipart==0.0.9
starlette==0.36.2
pytest==7.4.0
pytest-asyncio==0.21.1 